Extracts class and getStatus method
